Massive Changes
As per usual, I made too many changes before commiting; too many to list.
That said, in general most everything writes when Dropped.
The block manager and root directory (I think) are both passed around as
Rc<RefCell<_>>, which I'm not thrilled about.

Testing could be better, and there are bugs.
It seems like it's something to do with overwriting files that exist.
Not sure really.
Overall, it's _ok_, but could be much better.

However, I tihnk the point needs to be to focus on expanding to support
programs running inside the FS.  This will demo well.