Bug squasing...
There were some fairly serious bugs in the last commit.
Two of the biggest were reading and wrining files.
I wasn't holding on to the reader/writer between FUSE calls.
This caused a bunch of disperate, chunks of  blocks to be created.
The end result was a file that only contained the n mod 65k bytes, when writing.
Reading it would return those last bytes of the file over and over.
And worst, it would never actually return the expected number of bytes.
Also, when a file was closed, the reader/writer wasn't benig disassociated from the inode.
So you could read the file, but only once, for example.