Two Things
First I renamed BlockStorage::metadata* to BlockStorage::map*.
Calling the map metadata was confusing along side the file/directory metadata.

Secondly, I fixed a crazy bug in the block-server.
It would not commit the block map properly when used with remote FUSE.
It turns out that the FileStore was getting dropped, and thus comitting the block map.

So now, we don't commit on drop, but surface a method: BlockStorage::commit_map that the BlockManager uses at the appropriate time.