First crypto!
Currently there is a master password for the file system. Disk blocks
are encrypted with a hashed password, and a nonce, using
XChaChha20.. The nonce is derived from the file and file version
UUID.

There's tons of functionality missing, most notably there is only one
password.  However, even worse is that each block is encrypted with
the same block. This needs to get dealt with ASAP.