Lower Level Encryption
Added encryption to the bootstrapping readers and writers. These are
at a lower level than the BlockManager, and serve to read and write
the BlockMap.

Now that both the bootstrapping reading and writers, as well as the
readers and writers in BlockManager are both encrypting and de-
crypting, we should have pretty good coverage. Although, I'm thinking
about adding another layer of master-password encryption below
the BlockManager...