Utilize User
I'm now actually using the user when mounting a file system. Basically,
all access requires a master password. This is entered at a prompt
from the BlockServer when running that way. For local bundle access
it's entered when mounted.

Addititonally, when mounting, a user name and password is entered,
which are used for the current session.

A user name and password are entered when creating a file system,
and that user is added to the user's list.

When adding users, the master password, as well as an existing user
and password are all entered. There should probably be some level
of permissions that allows only certain people to add users.