Big, Nasty bug
There was a bug in the read_file implementaiton in fsimpl. I assumed
that blocks would be full. While adding a disassembler as a WASM
program, I was having weird problems writing to a file. It turns out
that each write would consume a block (yes, this is bogus). So when
reading bolcks back, my assumption caused the decryption to be
off, and everything got weird.