GET and POST are operational*
I've updated the permissions system to test for get and post grants.
The return value from the Wasm function is sent back to the HTTP server,
and forwarded to the client.
Event registration for Wasm programs has been completely reworked. Now
the events are only sent to programs that register, whereas before they
were sent to every program, and it was filtered prior to invoking the Wasm
function.

* There is a caveat: both calls block waiting for a result.