Integration Test!!
I've now got an integration test that tests the RESTful interface to Wasm
programs. It has two tests, and two file systems. One file system is
empty, and we test for failure to send REST messages.

The other file system has the echo wasm program, and we test GET and
POST for valid responses. The other RESTful methods don't have grants
to run, and we test for that condition.